+++ service/s3outposts/errors.go
@@ -26,6 +26,13 @@ const (
 	// There was an exception with the internal server.
 	ErrCodeInternalServerException = "InternalServerException"
 
+	// ErrCodeOutpostOfflineException for service response error code
+	// "OutpostOfflineException".
+	//
+	// The service link connection to your Outposts home Region is down. Check your
+	// connection and try again.
+	ErrCodeOutpostOfflineException = "OutpostOfflineException"
+
 	// ErrCodeResourceNotFoundException for service response error code
 	// "ResourceNotFoundException".
 	//
@@ -49,6 +56,7 @@ var exceptionFromCode = map[string]func(protocol.ResponseMetadata) error{
 	"AccessDeniedException":     newErrorAccessDeniedException,
 	"ConflictException":         newErrorConflictException,
 	"InternalServerException":   newErrorInternalServerException,
+	"OutpostOfflineException":   newErrorOutpostOfflineException,
 	"ResourceNotFoundException": newErrorResourceNotFoundException,
 	"ThrottlingException":       newErrorThrottlingException,
 	"ValidationException":       newErrorValidationException,